+++ src/puppetlabs/kitchensink/file.clj
@@ -1,11 +1,21 @@
 (ns puppetlabs.kitchensink.file
+  (:require [clojure.java.io :as io]
+            [clojure.string :as str])
   (:import (java.io BufferedWriter FileOutputStream OutputStreamWriter)
+           (java.nio ByteBuffer)
+           (java.nio.channels Channels)
            (java.nio.file CopyOption Files LinkOption Path Paths StandardCopyOption)
-           (java.nio.file.attribute FileAttribute PosixFilePermissions)))
+           (java.nio.file.attribute FileAttribute PosixFilePermissions)
+           (java.util.zip GZIPInputStream)
+           (org.apache.commons.compress.archivers.tar TarArchiveInputStream)))
 
 (defn str->path
   ^Path [path]
   (Paths/get path (into-array String [])))
+
+(defn dir+file->path
+  ^Path [directory file]
+  (Paths/get directory (into-array String [file])))
 
 (defn get-perms
   "Returns the currently set permissions of the given file path."
@@ -56,10 +66,12 @@
                          default-permissions))
          temp-attributes (into-array FileAttribute [(perms->attribute "rw-------")])
          temp-file (Files/createTempFile dir (.toString (.getFileName target)) "tmp" temp-attributes)
-         stream (proxy [FileOutputStream] [(.toString temp-file)]
-                  (close []
-                    (.sync (.getFD ^FileOutputStream this))
-                    (proxy-super close)))
+         stream  (proxy [FileOutputStream] [(.toString temp-file)]
+                   (close []
+                     (.sync (.getFD ^FileOutputStream this))
+                     ;; this looks weird, but makes the proxy-super avoid reflection by masking `this` with a version that has the meta tag
+                     (let [^FileOutputStream this this]
+                       (proxy-super close))))
          writer (BufferedWriter. (OutputStreamWriter. stream))]
 
      (write-function writer)
@@ -82,3 +94,47 @@
    (atomic-write-string path string nil))
   ([path string permissions]
    (atomic-write path #(.write ^BufferedWriter % ^String string) permissions)))
+
+(defn unzip-file
+  "Given a path to an input file, and a path to an output location, attempt to unzip the file.
+  Will not overwrite the same path with the output"
+  [input-file output-file]
+  (when (not= input-file output-file)
+    (io/make-parents output-file)
+    (with-open [file-input-stream (io/input-stream input-file)
+                gzip-input-stream (GZIPInputStream. file-input-stream)
+                file-output-stream (io/output-stream output-file)]
+      (io/copy gzip-input-stream file-output-stream))))
+
+(def empty-file-attributes
+  (into-array FileAttribute []))
+
+(defn write-tar-stream-to-file
+  [^TarArchiveInputStream input-stream ^Path output-path ^ByteBuffer buffer]
+  (with-open [out-channel (.getChannel (FileOutputStream. (.toFile output-path)))]
+    ;; move the byte buffer out for efficiency
+    (let [;; specifically don't close this as it will close the whole tarfile stream
+          in-channel (Channels/newChannel input-stream)]
+      (loop [in-buffer-size (.read in-channel buffer)]
+        (when (or (pos? in-buffer-size) (pos? (.position buffer)))
+          (.flip buffer)
+          (.write out-channel buffer)
+          (.compact buffer)
+          (recur (.read in-channel buffer)))))))
+
+(defn untar-file
+  "Given a path to a tar file, and a path to the parent directory to untar the repo in, recreate the contents of the tar file
+  Note: does not recreate original permissions, or support symlinks."
+  [path-to-tar-file output-directory]
+  (let [trimmed-output (if (str/ends-with? output-directory "/") (subs output-directory 0 (dec (count output-directory))) output-directory)
+        ;; allocate the copy buffer once and reuse for efficiency.
+        buffer (ByteBuffer/allocateDirect (* 64 1024))]
+    (with-open [tar-input-stream (TarArchiveInputStream. (io/input-stream path-to-tar-file))]
+      (loop [entry (.getNextTarEntry tar-input-stream)]
+        (when (some? entry)
+          (if (.isDirectory entry)
+            (Files/createDirectories (dir+file->path trimmed-output (.getName entry)) empty-file-attributes)
+            (let [output-file (dir+file->path trimmed-output (.getName entry))]
+              (io/make-parents output-file)
+              (write-tar-stream-to-file tar-input-stream output-file buffer)))
+          (recur (.getNextTarEntry tar-input-stream)))))))