+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 * change use of java.security.cert.X509Certificate/getSubjectDN, which is now deprecated, to java.security.cert.X509Certificate/getSubjectX500Principal
 * add function `get-lein-project-version` to retrieve a given project version from the standard `lein` system properties.
 * mark deprecated interfaces with metadata deprecation
+* add 'unzip-file' routine to the files namespace to decompress zipped files
+* add 'untar-file' routine to the files namespace to decompress tar files
+* update clj-parent to 5.6.6
+* remove reflection use in atomic-write
 
 ## 3.2.3
 * add clj-kondo linting, eastwood linting, and PR testing